separate into plots into multiple files that are easier to manage and extend
