+++ case_studies/sdss_galaxies/models/simulated_sdss_individual_galaxies.pt
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f1d665023892797832d807d68b6fc6ebb98d52000676fbe8ffb1525fe91a3220
+oid sha256:3d9723010d4e53f39d6eef564681087f6c1a930351ba9c267c4cf043891eba0a
 size 236236315
